Move test to ceparate file
