+++ .github/workflows/main.yml
@@ -33,6 +33,7 @@ jobs:
         xvfb-run -a py.test dioptas/tests/unit_tests
         xvfb-run -a py.test dioptas/tests/controller_tests/test_BatchController.py
         xvfb-run -a py.test dioptas/tests/controller_tests/test_BatchController_part2.py
+        xvfb-run -a py.test dioptas/tests/controller_tests/test_BatchController_part3.py
         xvfb-run -a py.test dioptas/tests/functional_tests/test_functional_integration.py
         xvfb-run -a py.test dioptas/tests/functional_tests/test_functional_JcpdsEditor.py
         xvfb-run -a py.test dioptas/tests/functional_tests/test_img_bg.py
